+++ frontend/src/App.tsx
@@ -1,6 +1,7 @@
 import { Navigate, Route, Routes } from 'react-router-dom'
 import { AppLayout } from './layouts/AppLayout'
 import { ApplicationDetailsPage } from './pages/ApplicationDetailsPage'
+import { ApplicationBoardPage } from './pages/ApplicationBoardPage'
 import { ApplicationsPage } from './pages/ApplicationsPage'
 import { DashboardPage } from './pages/DashboardPage'
 import { EditApplicationPage } from './pages/EditApplicationPage'
@@ -17,6 +18,7 @@ function App() {
         <Route element={<AppLayout />}>
           <Route index element={<Navigate to="/applications" replace />} />
           <Route path="dashboard" element={<DashboardPage />} />
+          <Route path="board" element={<ApplicationBoardPage />} />
           <Route path="applications" element={<ApplicationsPage />} />
           <Route path="applications/new" element={<NewApplicationPage />} />
           <Route path="applications/:id" element={<ApplicationDetailsPage />} />